MSPSDS-520: refactor image and document controllers, commonize them as much as possible
