--- web/app/helpers/addresses_helper.rb
@@ -1,2 +1,0 @@
-module AddressesHelper
-end
