--- web/app/views/documents/_table.html.slim
@@ -1,30 +1,0 @@
-table.govuk-table.document-list
-  thead.govuk-table__head
-    tr.govuk-table__row
-      th.govuk-table__header[scope="col" colspan="2"]
-  tbody.govuk-table__body
-    - @documents.each do |document|
-      tr.govuk-table__row
-        td.document-preview.govuk-table__cell
-          = link_to render("document_image", document: document), document, class: "document-link", target: "_blank"
-        td.document-details.govuk-table__cell
-          = link_to document do
-            h2.govuk-heading-m.document-title[class="govuk-!-margin-bottom-1"]
-              = document.metadata["title"]
-          - if document.metadata["updated"].present?
-            .document-updated.right
-              | Updated
-              '
-              = Time.zone.parse(document.metadata["updated"]).strftime("%d/%m/%Y")
-          .document-type-details[class="govuk-!-margin-bottom-3"]
-            span.document-type[class="govuk-!-font-weight-bold"]
-              = document.metadata["document_type"]&.titleize
-              = " – #{document.metadata['other_type'].titleize}" if document.metadata["other_type"].present?
-            span.document-filetype[class="govuk-!-padding-left-4"]
-              = document_filetype_label(document)
-          p.document-description[class="govuk-!-margin-bottom-1"]
-            = document.metadata["description"]
-          = link_to "Edit document details", edit_associated_document_path(@parent, document), \
-              class: "edit-document-link"
-          = link_to "Delete document", associated_document_path(@parent, document), method: :delete, \
-              data: { confirm: "Are you sure?" }, class: "delete-document-link"
