+++ web/app/views/documents_flow/upload.html.slim
@@ -2,6 +2,6 @@
 - content_for :after_header do
   = link_to "Back", @parent, class: "govuk-back-link"
 = render "page_heading", title: "New document"
-= form_with model: @document, scope: :document, url: wizard_path, method: :put, local: true do |form|
+= form_with scope: :document, url: wizard_path, method: :put, local: true do |form|
   = render "form_validation_display", errors: @errors
   = render "file_upload_display", form: form
