+++ web/app/views/images/edit.html.slim
@@ -2,4 +2,4 @@
 - content_for :after_header do
   = link_to "Back", @parent, class: "govuk-back-link"
 = render "page_heading", title: "Edit image details"
-= render "metadata_form", image: @image, show_delete: true, target_url: associated_image_path(@parent, @image)
+= render "metadata_form", image: @file, show_delete: true, target_url: associated_image_path(@parent, @file)
