+++ web/app/views/images_flow/metadata.html.slim
@@ -2,4 +2,4 @@
 - content_for :after_header do
   = link_to "Back", @parent, class: "govuk-back-link"
 = render "page_heading", title: "Enter image details"
-= render "images/metadata_form", image: @image_blob, show_delete: false, target_url: wizard_path
+= render "images/metadata_form", image: @file_blob, show_delete: false, target_url: wizard_path
