+++ web/app/views/images_flow/upload.html.slim
@@ -2,6 +2,6 @@
 - content_for :after_header do
   = link_to "Back", @parent, class: "govuk-back-link"
 = render "page_heading", title: "New image"
-= form_with model: @image, scope: :image, url: wizard_path, method: :put, local: true do |form|
+= form_with scope: :image, url: wizard_path, method: :put, local: true do |form|
   = render "form_validation_display", errors: @errors
   = render "file_upload_display", form: form
